pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-showcase"
version = "0.1.0"
description = "A minimal showcase of LangGraph pipeline patterns"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langchain-mistralai>=0.2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "langsmith>=0.1.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
showcase = "showcase.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
]
ignore = [
    "E501",   # Line too long (handled by formatter)
]
